requirements.txt
GitPython
langchain-text-splitters
langchain-community
langgraph
langchain-chroma
chromadb
openai
groq
langchain-openai
langchain-cohere
langchain_huggingface
llama-cpp-python
openai
python-dotenv
duckduckgo_search
datasets
wikipedia
langchain_experimental
google-api-python-client
gpt4all